requirements.txt
pandas
numpy
scikit-learn
matplotlib
mlflow
notebook
tqdm
joblib

# Config & validation
python-box
pyyaml
ensure
types-PyYAML

# Serving
flask
flask-cors